+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v2.3.24
 
+- Added helper script `logout.sh` to call `hyprshutdown`
 - Moved layer rules to own file `LayerRules.conf`
   - Added additional rules from ``TheAhumMaitra``
   - Updated LUA config accordingly